Fix local var type

+++ nautilus_trader/cache/cache.pyx
@@ -243,7 +243,7 @@ cdef class Cache(CacheFacade):
         # As there should be a bi-directional one-to-one relationship between
         # caches and indexes, each cache and index must be checked individually
 
-        cdef double timestamp_us = unix_timestamp_us()
+        cdef int64_t timestamp_us = unix_timestamp_us()
         self._log.info("Checking data integrity...")
 
         # Needed type defs
